+++ src/index.ts
@@ -1,4 +1,4 @@
-export { default as StringUtils } from './utils/StringUtils';
-export { default as RandomUtils } from './utils/RandomUtils';
+export { default as BrazilianUtils } from './utils/BrazilianUtils';
 export { default as FormatUtils } from './utils/FormatUtils';
-export { default as ValidatorUtils } from './utils/ValidatorUtils';
+export { default as RandomUtils } from './utils/RandomUtils';
+export { default as StringUtils } from './utils/StringUtils';